Develop

+++ src/app/pages/user-profile/user-profile.component.ts
@@ -2,6 +2,7 @@ import { Component, inject } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { User } from 'src/app/interfaces/user.interface';
 import { UserService } from 'src/app/services/users.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-profile',
@@ -22,12 +23,28 @@ export class UserProfileComponent {
   }
 
   async deleteUser(id: string): Promise<void> {
-    alert('¿Quieres borrar este perfil de usuario?')
-    let response = await this.userService.delete(id);
-    console.log(response);
-    if (response) {
-      alert('El usuario ha sido borrado correctamente');
-      this.router.navigate(['/home'])
+    const result = await Swal.fire({
+      title: '¿Quieres eliminar este usuario?',
+      text: "Esta acción no podrá deshacerse",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#5a9de6',
+      cancelButtonColor: '#eb9b9b',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar'
+    });
+
+    if (result.isConfirmed) {
+      let response = await this.userService.delete(id);
+      console.log(response);
+      if (response) {
+        Swal.fire(
+          'Eliminado',
+          'El usuario se ha eliminado correctamente.',
+          'success'
+        );
+        this.router.navigate(['/home']);
+      }
     }
   }
 }
